Refactor #574: admin account / moderation 系を handler_stubs から分離
#573 (handler_stubs 全体トラッカー) の sub-issue #574 (Admin Account &
Moderation 系、★★★) の対応。調査の結果、scope の 11 endpoint は **全て
本実装済**であり、handler_stubs.go の名前と裏腹に real な handler だった
(Explore 調査時点の予想通り)。

handler_stubs.go ファイル名を最終的に削除する方針 (#573) に沿って、本
PR では account / moderation 系の 7 関数とそのテストを thematic file に
分離する。残りの endpoint は他 sub-issue (#575-#579) でそれぞれの担当
範囲を移して、handler_stubs.go / handler_stubs_test.go を最終的に空に
する。

主な変更点:

- internal/api/admin/accounts.go (新規): AccountsDelete /
  AccountsFindByEmail / DeleteAccount / scheduleAccountCascade を移動
- internal/api/admin/moderation.go (新規): UnsetUserAvatar /
  UnsetUserBanner / UpdateUserNote / DeleteAllFilesOfUser を移動
- internal/api/admin/accounts_test.go: 既存 (DeleteAllFilesOfUser /
  UpdateProxyAccount) に加えて Account 系テスト 12 件を追記
- internal/api/admin/moderation_test.go (新規): UnsetUserAvatar /
  UnsetUserBanner / UpdateUserNote のテスト
- internal/api/admin/handler_stubs.go: 移動先のメモコメントだけ残して
  該当関数を削除
- internal/api/admin/handler_stubs_test.go: 移動した関数のテストを
  削除 (約 192 行縮小)
- internal/testutil/mock_repository.go: applyProfileFields に
  moderationNote case を追加 (mock fidelity 改善)

handler_stubs_test.go のテスト強化:
- TestUnsetUserAvatar / Banner / UpdateUserNote が「204 を返す」だけ
  しか検証していなかったので、実際に repository の該当 field が
  クリアされる / 書き込まれることを検証する subtest を追加

#574 完了条件への到達状況:
- [x] 全 11 endpoint が本実装に置き換わる (調査の結果、最初から本実装済)
- [x] 既実装と判明したものは stub 経路 (handler_stubs.go) から外して
  thematic file に移動
- [x] 単体テスト追加 (3 endpoint で actual DB 更新 verify を追加)
- [-] e2e で admin ユーザーが上記操作を実行できることを確認
  (既存 e2e は変更なし、本 PR は handler 振る舞いを保ったままの
  refactor なので回帰リスクは低い)

Closes #574
